+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to
 ## Unreleased
 
 * Polysemy tracing added.
+* Polysemy input added.
+
 ## [0.9] - 2021-10-27
 
 * Remove extra quotes.